refactor(coding): relocate non-ui interaction logic
