+++ Makefile
@@ -53,6 +53,13 @@ HARNESSTUI_CODING_ADAPTERS := \
 	src/loushang/coding/ui/tool_blocks.py \
 	src/loushang/coding/ui/transcript_projection.py \
 	src/loushang/coding/ui/transcript_source.py
+CODING_TUI_PRODUCT_SOURCES := \
+	src/loushang/coding/model_selection.py \
+	src/loushang/coding/diagnostics/debug_status.py \
+	src/loushang/coding/event/presentation_policy.py \
+	src/loushang/coding/interaction/controller.py \
+	src/loushang/coding/interaction/intent.py \
+	src/loushang/coding/presentation/session.py
 HARNESSTUI_TEST_PATHS := \
 	tests/harnesstui \
 	tests/tui/test_clipboard_image.py \
@@ -94,6 +101,12 @@ HARNESSTUI_TEST_PATHS := \
 	tests/coding/test_ui_transcript_projection.py \
 	tests/coding/test_ui_transcript_source.py \
 	tests/coding/ui/test_screen_input.py
+CODING_TUI_PRODUCT_TEST_PATHS := \
+	tests/coding/test_ui_controller.py \
+	tests/coding/test_ui_debug_status.py \
+	tests/coding/test_ui_event_policy.py \
+	tests/coding/test_ui_model.py \
+	tests/coding/test_ui_session_view.py
 
 .PHONY: bootstrap test test-ai check-ai test-tui test-tui-render-contract lint-ai fmt-ai typecheck-ai typecheck-tui build-binary install-binary clean-binary vendor-ai-moonshot-anthropic-stream vendor-ai-moonshot-anthropic-complete vendor-ai-moonshot-anthropic-tools vendor-ai-moonshot-openai-stream vendor-ai-moonshot-openai-complete vendor-ai-moonshot-openai-tools vendor-ai-dashscope-openai-responses-stream vendor-ai-dashscope-openai-responses-tools example-ai-model-lookup example-ai-complete example-ai-stream example-ai-tools example-ai-typed-context example-ai-advanced-faux-stream example-ai-advanced-context-tools example-ai-advanced-tool-result-roundtrip example-ai-kimi-anthropic-stream example-ai-kimi-anthropic-complete example-ai-kimi-anthropic-tools example-ai-kimi-openai-stream example-ai-kimi-openai-complete example-ai-kimi-openai-tools example-ai-dashscope-openai-responses-stream example-ai-dashscope-openai-responses-tools example-ai-custom-base-url-openai-advanced example-ai-faux-stream example-ai-context-tools-minimal example-ai-tool-result-roundtrip
 .PHONY: check-ai-catalog check-ai-examples check-ai-imports check-ai-coverage
@@ -129,13 +142,13 @@ check-ai-coverage:
 check-harnesstui: lint-harnesstui typecheck-harnesstui test-harnesstui
 
 lint-harnesstui:
-	uv --cache-dir .uv-cache run --extra dev ruff check $(HARNESSTUI_SHARED_SOURCES) $(HARNESSTUI_CODING_ADAPTERS) src/loushang/coding/ui/status_provider.py $(HARNESSTUI_TEST_PATHS)
+	uv --cache-dir .uv-cache run --extra dev ruff check $(HARNESSTUI_SHARED_SOURCES) $(HARNESSTUI_CODING_ADAPTERS) $(CODING_TUI_PRODUCT_SOURCES) src/loushang/coding/ui/status_provider.py $(HARNESSTUI_TEST_PATHS) $(CODING_TUI_PRODUCT_TEST_PATHS)
 
 typecheck-harnesstui:
-	uv --cache-dir .uv-cache run --extra dev mypy --follow-imports=silent $(HARNESSTUI_SHARED_SOURCES) $(HARNESSTUI_CODING_ADAPTERS) src/loushang/coding/ui/status_provider.py
+	uv --cache-dir .uv-cache run --extra dev mypy --follow-imports=silent $(HARNESSTUI_SHARED_SOURCES) $(HARNESSTUI_CODING_ADAPTERS) $(CODING_TUI_PRODUCT_SOURCES) src/loushang/coding/ui/status_provider.py
 
 test-harnesstui:
-	uv --cache-dir .uv-cache run --extra dev pytest $(HARNESSTUI_TEST_PATHS) -m "not tui_render_contract" -q
+	uv --cache-dir .uv-cache run --extra dev pytest $(HARNESSTUI_TEST_PATHS) $(CODING_TUI_PRODUCT_TEST_PATHS) -m "not tui_render_contract" -q
 
 test-tui:
 	. .venv/bin/activate && python -m pytest tests/tui -q
