+++ src/loushang/coding/interaction/__init__.py
@@ -1,0 +1,1 @@
+"""Coding interaction grammar and control independent of UI surfaces."""
