+++ src/loushang/coding/presentation/__init__.py
@@ -1,0 +1,1 @@
+"""Coding product presentation adapters independent of UI surfaces."""
