+++ src/loushang/coding/prompt_command.py
@@ -6,7 +6,7 @@ import traceback
 from collections.abc import Mapping, Sequence
 from typing import Any, TextIO
 
-from loushang.coding.ui.model import ensure_usable_session_model
+from loushang.coding.model_selection import ensure_usable_session_model
 from loushang.coding.ui.plain_events import PlainCodingEventRenderer
 from loushang.coding.ui.plain_renderer import PlainCodingUiRenderer
 from loushang.coding.work_shell import CodingWorkShell
