+++ src/loushang/coding/ui/conversation_event_adapter.py
@@ -4,7 +4,7 @@ from collections.abc import Callable
 from dataclasses import dataclass
 from typing import Any
 
-from loushang.coding.ui.event_policy import is_cancelled_error_message
+from loushang.coding.event.presentation_policy import is_cancelled_error_message
 from loushang.coding.ui.tool_blocks import ToolTranscriptProjector
 from loushang.harnesstui.conversation.projection import ConversationProjector
