+++ src/loushang/coding/workflow/command.py
@@ -6,7 +6,7 @@ import traceback
 from pathlib import Path
 from typing import Any, TextIO
 
-from loushang.coding.ui.model import ensure_usable_session_model
+from loushang.coding.model_selection import ensure_usable_session_model
 from loushang.coding.workflow.fake_runtime import FakeWorkflowAdapter
 from loushang.coding.workflow.loader import load_workflow, resolve_workflow_files
 from loushang.coding.workflow.report import (
