+++ tests/coding/test_screen_coding_tui_loop.py
@@ -798,8 +798,8 @@ def test_screen_loop_abort_uses_first_pending_steer_before_running_steer() -> No
 
 
 def test_screen_loop_waits_for_abort_settle_before_running_popped_pending_steer() -> None:
+    from loushang.coding.interaction.controller import CodingUiController
     from loushang.coding.testing.tui.playback import ScreenTuiLoopPlayback
-    from loushang.coding.ui.controller import CodingUiController
     from loushang.coding.ui.mode import (
         _screen_abort_handler,
         _screen_prompt_handler,
@@ -832,8 +832,8 @@ def test_screen_loop_waits_for_abort_settle_before_running_popped_pending_steer(
 
 
 def test_screen_loop_dispatches_session_command_without_prompting_agent() -> None:
+    from loushang.coding.interaction.controller import CodingUiController
     from loushang.coding.testing.tui.playback import ScreenTuiLoopPlayback
-    from loushang.coding.ui.controller import CodingUiController
     from loushang.coding.ui.mode import _screen_prompt_handler
 
     playback = ScreenTuiLoopPlayback()
