+++ tests/coding/test_ui_debug_command.py
@@ -14,7 +14,7 @@ class _Renderer:
 
 def test_debug_command_handler_disables_debug_logging() -> None:
     from loushang.coding.ui.debug_command import DebugCommandHandler
-    from loushang.coding.ui.intent import DebugIntent
+    from loushang.coding.interaction.intent import DebugIntent
 
     renderer = _Renderer()
     emitted: list[str] = []
@@ -50,7 +50,7 @@ def test_debug_command_handler_disables_debug_logging() -> None:
 
 def test_debug_command_handler_enables_debug_logging_with_scopes() -> None:
     from loushang.coding.ui.debug_command import DebugCommandHandler
-    from loushang.coding.ui.intent import DebugIntent
+    from loushang.coding.interaction.intent import DebugIntent
 
     session = object()
     renderer = _Renderer()
