+++ tests/coding/test_ui_event_policy.py
@@ -4,7 +4,7 @@ from types import SimpleNamespace
 
 
 def test_event_policy_writes_only_visible_transcript_events() -> None:
-    from loushang.coding.ui.event_policy import event_writes_transcript
+    from loushang.coding.event.presentation_policy import event_writes_transcript
 
     assert event_writes_transcript({"type": "message_start", "message": SimpleNamespace(role="user")})
     assert not event_writes_transcript({"type": "message_start", "message": SimpleNamespace(role="assistant")})
@@ -17,14 +17,14 @@ def test_event_policy_writes_only_visible_transcript_events() -> None:
 
 
 def test_event_policy_skips_internal_state_events() -> None:
-    from loushang.coding.ui.event_policy import event_writes_transcript
+    from loushang.coding.event.presentation_policy import event_writes_transcript
 
     for event_type in ("agent_start", "turn_start", "turn_end", "queue_update", "message_update"):
         assert not event_writes_transcript({"type": event_type})
 
 
 def test_event_policy_writes_agent_end_only_for_visible_errors() -> None:
-    from loushang.coding.ui.event_policy import event_writes_transcript
+    from loushang.coding.event.presentation_policy import event_writes_transcript
 
     assert not event_writes_transcript(
         {
@@ -53,7 +53,7 @@ def test_event_policy_writes_agent_end_only_for_visible_errors() -> None:
 
 
 def test_cancelled_error_message_detection_matches_abort_variants() -> None:
-    from loushang.coding.ui.event_policy import is_cancelled_error_message
+    from loushang.coding.event.presentation_policy import is_cancelled_error_message
 
     assert is_cancelled_error_message("Request cancelled.")
     assert is_cancelled_error_message("Request aborted by user")
