+++ tests/coding/test_ui_model.py
@@ -30,20 +30,64 @@ class _Session:
         self.current = normalized
 
 
+class _ScopedSession:
+    def __init__(self, scoped_models: object) -> None:
+        self.scopedModels = scoped_models
+
+
 def test_model_label_from_selection_hides_unknown_model() -> None:
-    from loushang.coding.ui.model import model_label_from_selection
+    from loushang.coding.model_selection import model_label_from_selection
 
     assert model_label_from_selection(ModelSelection(provider="unknown", model_id="unknown")) is None
 
 
 def test_model_label_from_selection_formats_provider_and_model() -> None:
-    from loushang.coding.ui.model import model_label_from_selection
+    from loushang.coding.model_selection import model_label_from_selection
 
     assert model_label_from_selection(ModelSelection(provider="kimi-code", model_id="kimi-for-coding")) == "kimi-code/kimi-for-coding"
 
 
+def test_model_selection_normalization_strips_external_identifiers() -> None:
+    from loushang.coding.model_selection import normalize_model_selection
+
+    assert normalize_model_selection(
+        {
+            "provider": " kimi-code ",
+            "model_id": " kimi-for-coding ",
+            "endpoint_id": " anthropic ",
+        }
+    ) == ModelSelection(
+        provider="kimi-code",
+        model_id="kimi-for-coding",
+        endpoint_id="anthropic",
+    )
+
+
+def test_current_model_first_preserves_relative_order() -> None:
+    from loushang.coding.model_selection import current_model_first
+
+    items = ["provider/a", "provider/b", "provider/a"]
+
+    assert current_model_first(
+        items,
+        current_label="provider/a",
+        label_of=lambda item: item,
+    ) == ["provider/a", "provider/a", "provider/b"]
+
+
+def test_iter_scoped_model_selections_keeps_distinct_endpoints_and_dedupes() -> None:
+    from loushang.coding.model_selection import iter_scoped_model_selections
+
+    first = ModelSelection(provider="provider", endpoint_id="first", model_id="model")
+    duplicate = {"model": first}
+    second = ModelSelection(provider="provider", endpoint_id="second", model_id="model")
+    session = _ScopedSession([{"model": first}, duplicate, {"model": second}])
+
+    assert asyncio.run(iter_scoped_model_selections(session)) == [first, second]
+
+
 def test_ensure_usable_session_model_keeps_existing_usable_model() -> None:
-    from loushang.coding.ui.model import ensure_usable_session_model
+    from loushang.coding.model_selection import ensure_usable_session_model
 
     current = ModelSelection(provider="kimi-code", model_id="kimi-for-coding")
     session = _Session(current=current)
@@ -55,7 +99,7 @@ def test_ensure_usable_session_model_keeps_existing_usable_model() -> None:
 
 
 def test_ensure_usable_session_model_prefers_kimi_coding_anthropic_detail() -> None:
-    from loushang.coding.ui.model import ensure_usable_session_model
+    from loushang.coding.model_selection import ensure_usable_session_model
 
     preferred = Model(id="kimi-for-coding", provider="kimi-code", endpoint="kimi-code-anthropic")
     fallback = Model(id="kimi-for-coding", provider="kimi-code", endpoint="openai-completions:cn:coding")
@@ -71,7 +115,7 @@ def test_ensure_usable_session_model_prefers_kimi_coding_anthropic_detail() -> N
 
 
 def test_ensure_usable_session_model_falls_back_to_available_selection() -> None:
-    from loushang.coding.ui.model import ensure_usable_session_model
+    from loushang.coding.model_selection import ensure_usable_session_model
 
     fallback = ModelSelection(provider="kimi-code", model_id="kimi-for-coding")
     session = _Session(
