+++ tests/coding/test_ui_prompt_dispatch.py
@@ -35,8 +35,8 @@ class _Controller:
 
 
 def test_prompt_dispatch_begins_and_ends_work_intents() -> None:
-    from loushang.coding.ui.controller import ControllerResult
-    from loushang.coding.ui.intent import PromptIntent
+    from loushang.coding.interaction.controller import ControllerResult
+    from loushang.coding.interaction.intent import PromptIntent
     from loushang.coding.ui.prompt_dispatch import PromptDispatchHandler
 
     lifecycle = _Lifecycle()
@@ -66,8 +66,8 @@ def test_prompt_dispatch_begins_and_ends_work_intents() -> None:
 
 
 def test_prompt_dispatch_does_not_start_lifecycle_for_non_work_intents() -> None:
-    from loushang.coding.ui.controller import ControllerResult
-    from loushang.coding.ui.intent import QuitIntent
+    from loushang.coding.interaction.controller import ControllerResult
+    from loushang.coding.interaction.intent import QuitIntent
     from loushang.coding.ui.prompt_dispatch import PromptDispatchHandler
 
     lifecycle = _Lifecycle()
@@ -90,7 +90,7 @@ def test_prompt_dispatch_does_not_start_lifecycle_for_non_work_intents() -> None
 
 
 def test_prompt_dispatch_ends_work_when_controller_raises() -> None:
-    from loushang.coding.ui.intent import BashIntent
+    from loushang.coding.interaction.intent import BashIntent
     from loushang.coding.ui.prompt_dispatch import PromptDispatchHandler
 
     lifecycle = _Lifecycle()
